Add files via upload
This is the Group Project, For the topic "Design a database application using Python GUI to modify specified records of a residential society using flat no. using database and display the modified record."
(flat no, area of the flat owner, no. of family members)
Group Members are Nandini, Gayatri , Leena , Deepika and Anushka.